requirements.txt
discord.py>=2.3.0
motor>=3.3.2
python-dotenv>=1.0.0
pymongo>=4.5.0
